analyst/requirements.txt
fastapi
uvicorn[standard]
aioredis
sqlite3-binary; python_version < '3.10'  # otherwise sqlite3 builtin
python-dotenv
